requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.6
redis==5.0.1
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
ffmpeg-python==0.2.0
pillow==10.1.0
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
